// frontend/src/types/product.ts
export interface ProductWithDetails {
    product: Product;        // The product details
    artisan: Artisan;        // The artisan who created the product
    reviews: Review[];       // Array of reviews for the product
    relatedProducts: Product[]; // Array of related products
}


export interface Product {
    _id: string;                  // MongoDB Object ID for the product
    title: string;                // Title of the product
    images: string[];             // Array of image URLs for the product
    description: string;          // Product description
    materials: string[];          // Materials used for the product
    createdBy: string;            // The slug of the artisan (relates to Artisan type)
    category: string;             // Category of the product (e.g., "Art", "Craft")
    price: number;                // Price of the product
    rating: number;               // Rating of the product
    slug: string;                 // Slug for the product (unique identifier)
    createdAt: string;            // Timestamp of when the product was created
    updatedAt: string;            // Timestamp of when the product was last updated
    __v: number;                  // Version key (auto-generated by MongoDB)
}

export interface Artisan {
    _id: string;                  // MongoDB Object ID for Artisan
    fullName: string;              // Full name of the artisan
    description: string;          // Artisan's description
    profilePic: string;           // URL to the artisan's profile picture
    slug: string;                 // Artisan's slug (unique identifier)
    createdAt: string;            // Timestamp of when the artisan was created
    updatedAt: string;            // Timestamp of when the artisan was last updated
    __v: number;                  // Version key (auto-generated by MongoDB)
}

export interface Review {
    _id: string;                  // MongoDB Object ID for the review
    productId: string;            // The ID of the product this review is related to
    user: string;
    rating: number;               // Rating given in the review
    comment: string;              // Text of the review comment
    createdAt: string;            // Timestamp of when the review was created
    updatedAt: string;
}


export interface RelatedProduct {
    relatedProducts: Product[];
}